refactor: HTTP 에 관한 정보를 가지는 객체 분리
